remove helpers.ts
